Refactor getConstituencyData func out so it can be re-used in other components
